+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Unreleased
+## 0.20.0 (2025-01-07)
 
 ### titiler.core